+++ lib/src/backends/local/local_file_system.dart
@@ -17,6 +17,9 @@ class LocalFileSystem extends FileSystem {
   @override
   Link link(path) => new _LocalLink(this, shim.newLink(path));
 
+  @override
+  String get pathSeparator => shim.pathSeparator;
+
   /// Gets the directory provided by the operating system for creating temporary
   /// files and directories in. The location of the system temp directory is
   /// platform-dependent, and may be set by an environment variable.