+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `FileSystem.systemTempDirectory`
 * Added the ability to pass `Uri` and `FileSystemEntity` types to
   `FileSystem.directory()`, `FileSystem.file()`, and `FileSystem.link()`
+* Added `FileSystem.pathSeparator`
 
 #### 1.0.0